+++ main.go
@@ -45,6 +45,10 @@ func main() {
 		err = runBattery()
 	case "diagnose":
 		err = runDiagnose()
+	case "debug":
+		err = runDebug(os.Args[2:])
+	case "buttons":
+		err = runButtons(os.Args[2:])
 	case "--daemon", "-d", "daemon":
 		err = runDaemon()
 	case "update":
@@ -89,6 +93,8 @@ Usage:
   jabridge status      show connected USB devices
   jabridge battery     show headset battery from 0 to 100 percent
   jabridge diagnose    check USB and control-interface discovery
+  jabridge debug       create a shareable device and service report
+  jabridge buttons     listen to headset media buttons and volume wheels
   jabridge update      update the app
   jabridge firmware    check or download device firmware
   jabridge settings    list or change supported device settings